Trying to write a serializer and start to figure out what is needed for enrollment report
Testing is a bear because of all the database constraints. Had to remove
the foreign key constraints, but still having errors creating actual
enrollments.

In the app/models/query/enrollments.rb there is the query pulled from
Olympia's SQL and some notes on the way that I wanted to serialize the
entities and also the way I declared models.